+++ src/index.js
@@ -2,7 +2,6 @@ import './assets/css/tailwind.css';
 
 import VueAdsTableContainer from './components/TableContainer';
 import VueAdsTable from './components/Table';
-import VueAdsAsyncTable from './components/AsyncTable';
 
 export default VueAdsTableContainer;
-export { VueAdsTable, VueAdsAsyncTable };
+export { VueAdsTable };